Created Story Triggers for Day 1
I have set up the three triggers for Tutorial, Maeve, and Final. These will interact with the player when they enter the trigger, and will destroy themselves when the player has finished interacting with them. We still need the player dialog to put in place, and a few adjustments will have to be made after playtesting.